feat: add auth to all the api endpoints

+++ src/app/api/seed/route.tsx
@@ -1,7 +1,7 @@
 import { Cylinder } from '@/lib/models/cylinder'
 import { Fill } from '@/lib/models/fill'
 import { Maintenance } from '@/lib/models/maintenance'
-
+import { auth } from '@/auth'
 import { Client } from '@/lib/models/client'
 import dayjs from 'dayjs'
 
@@ -167,6 +167,13 @@ function generateMaintenance(): Maintenance[] {
 }
 
 export async function GET(request: Request) {
+	const session = await auth()
+	if (!session)
+		return Response.json(
+			{ error: 'auth', message: 'Must be logged in' },
+			{ status: 401 },
+		)
+
 	if (process.env.NODE_ENV !== 'development') {
 		await sequelize.sync({ force: true })
 		return Response.json(
